Module Processor.Topology

CPU Topology.

The topology is static and built uppon Module load.

See https://github.com/haesbaert/ocaml-processor for a tutorial.

val t : Cpu.t list

t is the list of all logical cpus in the system.